requirements.txt
streamlit>=1.35.0
databricks-sdk>=0.20.0
pandas>=1.5.0
